Start of dcdr-related refactoring (Converting project into library)

+++ external_libraries/common_src/vboindexer.cpp
@@ -102,7 +102,7 @@ void indexVBO(
 	std::vector<glm::vec2> & in_uvs,
 	std::vector<glm::vec3> & in_normals,
 
-	std::vector<unsigned int> & out_indices,
+	std::vector<size_t> & out_indices,
 	std::vector<glm::vec3> & out_vertices,
 	std::vector<glm::vec2> & out_uvs,
 	std::vector<glm::vec3> & out_normals
